Fix the I/O of Mesh_complex_3_in_triangulation_3, when the index types are   not all "int".
LR
part of r61189 from Mesh_3-experimental-GF

+++ Mesh_3/include/CGAL/Mesh_3/Mesh_surface_cell_base_3.h
@@ -241,7 +241,7 @@ operator<<(std::ostream &os,
     if(is_ascii(os))
       os << ' ' << c.surface_patch_index(i);
     else
-      write(os, static_cast<int>(c.surface_patch_index(i)));
+      write(os, c.surface_patch_index(i));
   }
   return os;
 }
